requirements.txt
future
faker
motor
PyDictionary==2.0.1
emoji
lxml
beautifulsoup4
requests
python-telegram-bot==12.8
sqlalchemy==1.3.20
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
psutil
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon
spamwatch
alphabet_detector
cachetools
fontTools
gtts
tswift
pytz
youtube_dl
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
pyrogram==1.4.12
datetime
aiohttp>=3.7.4
asyncio
pykeyboard==0.1.2
Pillow>=4.2.0
CurrencyConverter
googletrans==4.0.0-rc1
google_trans_new
pyrate-limiter
nekos.py
tgcrypto
python-arq==6.0.3
hachoir
configparser
bing_image_downloader
parsel
urllib3
heroku3
telegraph
youtube_search
search_engine_parser
glitch_this
NumPy
opencv-python-headless
countryinfo==0.1.2
secureme #encrypt
aiofiles
aiohttp>=2.2.5
wget
pymongo==3.11.0
dnspython
psutil==5.8.0
gpytranslate
envparse
flask==2.1.1
werkzeug==2.0.3
flag
yt_dlp
pendulum
